Fix detection of java for flatpak
The logic for fdroidserver to find java is defined here:
https://gitlab.com/fdroid/fdroidserver/blob/0.8/fdroidserver/common.py#L154

For me it was not possible to match the regex as I wasn't able to move
the directory of java in flatpak. All I can do are symlinks, but the
fdroidserver code checks whether a given JAVA_HOME is a symlink and then
drops that path.

This is the best solution I came up with, which works now. In future, I
want to update fdroidserver directly to allow flatpak installations in
its regex, but as we don't want to bump fdroidserver before 1.0.0, I
decided to go this way.

+++ repomaker/__init__.py
@@ -69,23 +69,32 @@ def requirements_check(app_configs, **kwargs):  # pylint: disable=unused-argumen
     config = {}
     common.fill_config_defaults(config)
     common.config = config
+    java_home_in_flatpak = '/usr/lib/sdk/openjdk9'
     if 'keytool' not in config:
-        errors.append(
-            Error(
-                'Could not find `keytool` program.',
-                hint='This program usually comes with Java. Try to install JRE. '
-                     'On Debian-based system you can try to run '
-                     '`apt install openjdk-8-jre-headless`.',
+        if os.path.exists(java_home_in_flatpak):
+            config['keytool'] = os.path.join(java_home_in_flatpak, 'bin', 'keytool')
+            common.config = config
+        else:
+            errors.append(
+                Error(
+                    'Could not find `keytool` program.',
+                    hint='This program usually comes with Java. Try to install JRE. '
+                         'On Debian-based system you can try to run '
+                         '`apt install openjdk-8-jre-headless`.',
+                )
             )
-        )
     if 'jarsigner' not in config and not common.set_command_in_config('apksigner'):
-        errors.append(
-            Error(
-                'Could not find `jarsigner` or `apksigner`. At least one of them is required.',
-                hint='Please install the missing tool. On Debian-based systems you can try to run '
-                     '`apt install apksigner`.',
+        if os.path.exists(java_home_in_flatpak):
+            config['jarsigner'] = os.path.join(java_home_in_flatpak, 'bin', 'jarsigner')
+            common.config = config
+        else:
+            errors.append(
+                Error(
+                    'Could not find `jarsigner` or `apksigner`. At least one of them is required.',
+                    hint='Please install the missing tool. On Debian-based systems '
+                         'you can try to run `apt install apksigner`.',
+                )
             )
-        )
     # aapt
     try:
         common.SdkToolsPopen(['aapt', 'version'], output=False)
